Fix javac error again.
